unit tests added by max and aisha

+++ speed_test/speed_test/src/main/java/com/example/speed_test/SpeedRepository.java
@@ -3,6 +3,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface SpeedRepository extends JpaRepository<SpeedResult, Double>{
+public interface SpeedRepository extends JpaRepository<SpeedResult, Integer>{
 
 }
